pyproject.toml
[project]
name = "BuildAnAIAgentInPython"
version = "0.1.0"
description = "Boot.dev course: Build an AI agent in Python"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai==1.12.1",
    "google-generativeai>=0.8.5",
    "python-dotenv==1.1.0",
]
